# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Documentation",
        "description": "Test project documentation from VK",
        "contact": {
            "email": "desk10567@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://0.0.0.0:8080",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/comment/{comment_id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Index",
                "description": "Get a comment.",
                "operationId": "commentIndex",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Authorization token",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer 67pIVg4vAbmRZewXK4H0llU4XQtSy3VIxTpofgUOFfpcTfXfK1U0tSM3YNhytEw7BA7KzePrRJoQs3KlQSpzCer10lObTrep8wuZ"
                        }
                    },
                    {
                        "name": "comment_id",
                        "in": "path",
                        "description": "comment id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IndexCommentResourceTrue"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResourceFalse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundCommentResourceTrue"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "IndexCommentResourceTrue": {
                "title": "IndexCommentResourceTrue",
                "description": "Index comment",
                "properties": {
                    "success": {
                        "title": "Success",
                        "description": "Status answer",
                        "type": "boolean",
                        "default": "true"
                    },
                    "message": {
                        "title": "Message",
                        "description": "Response message",
                        "type": "string",
                        "default": "comment found"
                    },
                    "data": {
                        "title": "Data",
                        "description": "Response data",
                        "type": "object",
                        "example": {
                            "id": 1,
                            "text": "long text ...",
                            "user_id": 2,
                            "parent_comment_id": null,
                            "date": "2023-05-10"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "IndexCommentResourceTrue"
                }
            },
...
            "LoginResourceFalse": {
                "title": "LoginResourceFalse",
                "description": "Login resource",
                "properties": {
                    "success": {
                        "title": "Success",
                        "description": "Status answer",
                        "type": "boolean",
                        "default": "false"
                    },
                    "message": {
                        "title": "Message",
                        "description": "Response message",
                        "type": "string",
                        "default": "unauthorized"
                    },
                    "data": {
                        "title": "Data",
                        "description": "Response data",
                        "type": "object",
                        "default": []
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LoginResourceFalse"
                }
            },
...
            "NotFoundCommentResourceTrue": {
                "title": "NotFoundCommentResourceTrue",
                "description": "Not found comment",
                "properties": {
                    "success": {
                        "title": "Success",
                        "description": "Status answer",
                        "type": "boolean",
                        "default": "false"
                    },
                    "message": {
                        "title": "Message",
                        "description": "Response message",
                        "type": "string",
                        "default": "not found"
                    },
                    "data": {
                        "title": "Data",
                        "description": "Response data",
                        "type": "object",
                        "default": null
                    }
                },
                "type": "object",
                "xml": {
                    "name": "NotFoundCommentResourceTrue"
                }
            },
...
        }
    },
...
}